// src/mocks/types.ts
export type ProfileTabCommonType = {
    icon: string
    value: string
    property: string
}
export type ProfileTeamsType = ProfileTabCommonType & { color: ThemeColor }